Moved versioning plugin to specific package
