+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.versioning.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.VersioningPlugin
+implementation-class=org.shipkit.internal.gradle.version.VersioningPlugin
