+++ src/test/groovy/org/shipkit/internal/gradle/VersioningPluginTest.groovy
@@ -4,6 +4,7 @@ import org.shipkit.gradle.BumpVersionFileTask
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin
 import org.shipkit.internal.gradle.git.GitCommitTask
 import org.shipkit.internal.gradle.git.GitPlugin
+import org.shipkit.internal.gradle.version.VersioningPlugin
 import org.shipkit.internal.version.VersionInfo
 import testutil.PluginSpecification
