+++ src/test/groovy/org/shipkit/internal/gradle/java/ComparePublicationsPluginTest.groovy
@@ -6,7 +6,7 @@ import org.shipkit.gradle.ShipkitConfiguration
 import org.shipkit.gradle.java.ComparePublicationsTask
 import org.shipkit.gradle.java.DownloadPreviousPublicationsTask
 import org.shipkit.internal.gradle.ShipkitBintrayPlugin
-import org.shipkit.internal.gradle.VersioningPlugin
+import org.shipkit.internal.gradle.version.VersioningPlugin
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin
 import testutil.PluginSpecification
